basic timer functionality improved
