+++ src/client/api.ts
@@ -4,7 +4,8 @@ import { createHash, hmacHex } from '../crypto'
 import _ from 'lodash'
 import { isObject } from '../utils'
 import config from '../../configuration'
-import aoStore, { Task, Grid } from './store'
+import aoStore from './store'
+import { Task, Grid } from '../interfaces'
 import { io } from 'socket.io-client'
 
 import { runInAction, reaction } from 'mobx'
