+++ src/components/completed.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { computed, makeObservable } from 'mobx'
 import { observer } from 'mobx-react'
-import aoStore, { Task } from '../client/store'
+import aoStore from '../client/store'
+import { Task } from '../interfaces'
 import api from '../client/api'
 import AoStack from './stack'
 import { CardPlay } from '../cardTypes'
