+++ src/components/deck.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
 import { computed, makeObservable } from 'mobx'
 import { observer } from 'mobx-react'
-import aoStore, { Task, emptySearchResults } from '../client/store'
+import aoStore, { emptySearchResults } from '../client/store'
+import { Task } from '../interfaces'
 import InfiniteScroll from 'react-infinite-scroller'
 import AoContextCard from './contextCard'
 import AoDragZone from './dragZone'
