+++ src/components/dock.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
 import { computed, comparer, reaction, observable } from 'mobx'
 import { observer, Observer } from 'mobx-react'
-import aoStore, { Task } from '../client/store'
+import aoStore from '../client/store'
+import { Task } from '../interfaces'
 import api from '../client/api'
 import AoHome from './home'
 import AoHopper from './hopper'
