+++ src/components/draw.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { observer } from 'mobx-react'
 import { computed, makeObservable } from 'mobx'
-import aoStore, { Task } from '../client/store'
+import aoStore from '../client/store'
+import { Task } from '../interfaces'
 import { goInCard, findOrphans, findFirstCardInCard } from '../cardTypes'
 import BuddhaDoge from '../assets/images/buddadoge.svg'
 import RedBoat from '../assets/images/boatbtnselected.svg'
