+++ src/components/gem.tsx
@@ -6,7 +6,8 @@ import AoCardComposer from './cardComposer'
 import api from '../client/api'
 import { goInCard } from '../cardTypes'
 import Tippy from '@tippyjs/react'
-import aoStore, { Task } from '../client/store'
+import aoStore from '../client/store'
+import { Task } from '../interfaces'
 import Quill from '../assets/images/compose.svg'
 import 'tippy.js/dist/tippy.css'
 import 'tippy.js/themes/translucent.css'
