+++ src/components/members.tsx
@@ -2,7 +2,8 @@ import * as React from 'react'
 import { computed, makeObservable } from 'mobx'
 import { observer } from 'mobx-react'
 import InfiniteScroll from 'react-infinite-scroller'
-import aoStore, { Task } from '../client/store'
+import aoStore from '../client/store'
+import { Task } from '../interfaces'
 import api from '../client/api'
 import AoTip from './tip'
 import AoContextCard from './contextCard'
