+++ src/components/missions.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
 import { computed, makeObservable } from 'mobx'
 import { observer } from 'mobx-react'
-import aoStore, { Task } from '../client/store'
+import aoStore from '../client/store'
+import { Task } from '../interfaces'
 import AoStack from './stack'
 import { gloss } from '../semantics'
 import AoPopupPanel from './popupPanel'
