+++ src/components/proposals.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react'
 import { computed, makeObservable } from 'mobx'
 import { observer } from 'mobx-react'
-import aoStore, { Task, Signature } from '../client/store'
+import aoStore from '../client/store'
+import { Task, Signature } from '../interfaces'
 import AoStack from './stack'
 import { mostRecentSignaturesOnly, countCurrentSignatures } from '../cards.js'
 import AoTip from './tip'
