+++ src/components/reminder.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react'
 import { observer } from 'mobx-react'
-import aoStore, { Task } from '../client/store'
+import aoStore from '../client/store'
+import { Task } from '../interfaces'
 import api from '../client/api'
 import Gift from '../assets/images/gift.svg'
 import Tippy from '@tippyjs/react'
