+++ src/lib/actions.ts
@@ -12,28 +12,3 @@ export function taskCreate(
     inId
   }
 }
-
-// export function gridCreate(name: number, length: number, width: number) {
-//   return { type: 'grid-created', name, length, width }
-// }
-
-// export function gridUnpin(taskId: string, x: number, y: number) {
-//   return { type: 'grid-unpin', gridId, x, y }
-// }
-// export function gridPin(inId: string, taskId: string, x: number, y: number) {
-//   return { type: 'grid-pin', gridId, taskId, x, y }
-// }
-// export function gridResize(
-//   gridId: string,
-//   amount: number,
-//   side: 'right' | 'left' | 'top' | 'bottom'
-// ) {
-//   return { type: 'grid-resize', gridId, amount, side }
-// }
-
-// export type AoAction =
-//   | ReturnType<typeof taskCreate>
-//   | ReturnType<typeof gridCreate>
-//   | ReturnType<typeof gridResize>
-//   | ReturnType<typeof gridPin>
-//   | ReturnType<typeof gridUnpin>
